Use TCP socket instead of HTTP for communication
Use a simple TCP socket for communication between client and server
instead of HTTP. This speeds up processing.

Adds an (OSM) id to the protocol between client and server which is used
for error messages. Unfortunately this adds another parameter to several
functions which means the tests all have to add this parameter, too.